% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM.R
\name{LCM}
\alias{LCM}
\title{Least Cost Method}
\usage{
LCM(ex_matrix)
}
\arguments{
\item{ex_matrix}{A cost matrix where last column must be the supply and last row must be the demand. Input matrix should not have any missing values (NA), otherwise function will throw an error. It should be balanced i.e. total demand must be equal to total supply.}
}
\value{
A List which contain the allocation matrix and the total optimized cost.
}
\description{
The main objective is to minimize the total transportation cost, transport as much as possible through those routes (cells) where the unit transportation cost is lowest. This method takes into account the minimum cost of transportation for obtaining the initial solution.
}
\examples{
#Input matrix where last row is the Demand and last column is the Supply
ex_matrix=data.frame(D1=c(6,3,4,20),E1=c(4,8,4,95),F1=c(1,7,2,35),
                     Supply=c(50,40,60,150),row.names = c("A1","B1","C1","Demand"))
LCM(ex_matrix)

}
